+++ crypto_kem/sntrup653/clean/crypto_encode_653x4621.c
@@ -15,8 +15,8 @@ void PQCLEAN_SNTRUP653_CLEAN_crypto_encode_653x4621(unsigned char *out, const vo
     uint32 r2;
 
     for (i = 0; i < 326; ++i) {
-        r0 = (R0[2 * i] + 2310) & 16383;
-        r1 = (R0[2 * i + 1] + 2310) & 16383;
+        r0 = (uint16) ((R0[2 * i] + 2310) & 16383);
+        r1 = (uint16) ((R0[2 * i + 1] + 2310) & 16383);
         r2 = r0 + r1 * (uint32)4621;
         *out++ = (unsigned char) r2;
         r2 >>= 8;
@@ -24,7 +24,7 @@ void PQCLEAN_SNTRUP653_CLEAN_crypto_encode_653x4621(unsigned char *out, const vo
         r2 >>= 8;
         R[i] = (uint16) r2;
     }
-    R[326] = (R0[652] + 2310) & 16383;
+    R[326] = (uint16) ((R0[652] + 2310) & 16383);
 
     for (i = 0; i < 163; ++i) {
         r0 = R[2 * i];
@@ -123,5 +123,5 @@ void PQCLEAN_SNTRUP653_CLEAN_crypto_encode_653x4621(unsigned char *out, const vo
     R[0] = (uint16) r2;
 
     r0 = R[0];
-    *out++ = (unsigned char) r0; /*clang-analyzer-deadcode.DeadStores*/ /*r0 >>= 8;*/
+    *out++ = (unsigned char) r0;
 }
