+++ crypto_kem/sntrup761/clean/crypto_encode_761x4591.c
@@ -15,8 +15,8 @@ void PQCLEAN_SNTRUP761_CLEAN_crypto_encode_761x4591(unsigned char *out, const vo
     uint32 r2;
 
     for (i = 0; i < 380; ++i) {
-        r0 = (R0[2 * i] + 2295) & 16383;
-        r1 = (R0[2 * i + 1] + 2295) & 16383;
+        r0 = (uint16) ((R0[2 * i] + 2295) & 16383);
+        r1 = (uint16) ((R0[2 * i + 1] + 2295) & 16383);
         r2 = r0 + r1 * (uint32)4591;
         *out++ = (unsigned char) r2;
         r2 >>= 8;
@@ -24,7 +24,7 @@ void PQCLEAN_SNTRUP761_CLEAN_crypto_encode_761x4591(unsigned char *out, const vo
         r2 >>= 8;
         R[i] = (uint16) r2;
     }
-    R[380] = (R0[760] + 2295) & 16383;
+    R[380] = (uint16) ((R0[760] + 2295) & 16383);
 
     for (i = 0; i < 190; ++i) {
         r0 = R[2 * i];
@@ -143,5 +143,5 @@ void PQCLEAN_SNTRUP761_CLEAN_crypto_encode_761x4591(unsigned char *out, const vo
     r0 = R[0];
     *out++ = (unsigned char) r0;
     r0 >>= 8;
-    *out++ = (unsigned char) r0; /*clang-analyzer-deadcode.DeadStores*/ /*r0 >>= 8;*/
+    *out++ = (unsigned char) r0;
 }
